Centralize paging utilities in utils.js and refactor related imports
